% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary_codes.R
\name{ar1_cor}
\alias{ar1_cor}
\title{Construct a AR1 correlation structure}
\usage{
ar1_cor(n, rho)
}
\arguments{
\item{n}{The number of time}

\item{rho}{The correlation coefficient}
}
\value{
A correlation matrix
}
\description{
This function can be used to create a AR1 correlation structure.
}
