% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_emultir.R
\name{bootstrap.multir.plus.en}
\alias{bootstrap.multir.plus.en}
\title{The bootstrapping results for the eMultiR estimate based on one sampling}
\usage{
bootstrap.multir.plus.en(y,x,trt_a,yy,x_reg,r,time,proportion,id,dist)
}
\arguments{
\item{y}{The vector of secondary outcomes}

\item{x}{The covariate matrix for the secondary model}

\item{trt_a}{Treatment assignment (eg., 0,1)}

\item{yy}{The main outcome in binary}

\item{x_reg}{The covariate matrix used for both PS and CM models}

\item{r}{Missing indicator for subjects}

\item{time}{The number of visits}

\item{proportion}{The proportion of the subjects having the secondary outcome}

\item{id}{The id of subjects}

\item{dist}{The distribution of the secondary outcomes (so far only allow "gaussian")}
}
\value{
The list of bootstrapping estimates
}
\description{
This function can be used to obtain the bootstrapping result for the eMultiR estimate
based on one sampling.
}
