% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_emultir.R
\name{bootstrap.multir.plus.en.se}
\alias{bootstrap.multir.plus.en.se}
\title{The bootstrapping standard deviation for the eMultiR estimate}
\usage{
bootstrap.multir.plus.en.se(y,x,trt_a,yy,x_reg,r,time,proportion,id,dist,num.boost=50)
}
\arguments{
\item{y}{The vector of secondary outcomes}

\item{x}{The covariate matrix for the secondary model}

\item{trt_a}{Treatment assignment (eg., 0,1)}

\item{yy}{The main outcome in binary}

\item{x_reg}{The covariate matrix used for both PS and CM models}

\item{r}{Missing indicator for subjects}

\item{time}{The number of visits}

\item{proportion}{The proportion of the subjects having the secondary outcome}

\item{id}{The id of subjects}

\item{dist}{The distribution of the secondary outcomes (so far only allow "gaussian")}

\item{num.boost}{The number of bootstrapping. The default is 50}
}
\value{
The bootstrapping standard deviation
}
\description{
This function can be used to obtain the bootstrapping standard deviation for the eMultiR estimate.
}
