% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr.est}
\alias{dr.est}
\title{DR estimate}
\usage{
dr.est(prob_fit,trt_ind,main_outcome,pseudo_outcome_obs,
pseudo_outcome_comb,pseudo_trt)
}
\arguments{
\item{prob_fit}{The fitted probability in the treatment assignment model}

\item{trt_ind}{Treatment assignment (eg., 0,1)}

\item{main_outcome}{The main outcome in binary}

\item{pseudo_outcome_obs}{Pseudo outcome fitted from the first step of the CM model}

\item{pseudo_outcome_comb}{Pseudo outcome combined from treated and control groups
in the CM model}

\item{pseudo_trt}{Pseudo treatment assignment combined}
}
\value{
A vector of estimates
}
\description{
This function can be used to obtain the doubly robust estimate.
}
