% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary_codes.R
\name{est.fct.or}
\alias{est.fct.or}
\title{Estimating function for the CM model}
\usage{
est.fct.or(eta,pseudo_y,trt=1,n)
}
\arguments{
\item{eta}{The CM estimate}

\item{pseudo_y}{The pseudo outcome for treated (1) or control group (0)}

\item{trt}{Treatment assignment corresponding to `pseudo_y` (either 0 or 1)}

\item{n}{The sample size}
}
\value{
A correlation matrix
}
\description{
This function can be used to create a estimating function for the CM model.
}
