% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{iptw.binary}
\alias{iptw.binary}
\title{IPTW estimate}
\usage{
iptw.binary(prob_fit,trt_ind,main_outcome)
}
\arguments{
\item{prob_fit}{The fitted probability in the treatment assignment model}

\item{trt_ind}{Treatment assignment}

\item{main_outcome}{The main outcome in binary}
}
\value{
A vector of estimates
}
\description{
This function can be used to obtain the IPTW estimate.
}
