% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emultir.R
\name{MinBo.one}
\alias{MinBo.one}
\title{Information borrow scores}
\usage{
MinBo.one(s.outcome,cov.mat,n,proportion,time,r,id,dist="gaussian")
}
\arguments{
\item{s.outcome}{The vector of secondary outcomes}

\item{cov.mat}{The covariate matrix}

\item{n}{The sample size}

\item{proportion}{The proportion of the subjects having the secondary outcome
(so far only allow the value one)}

\item{time}{The number of visits}

\item{r}{Missing indicator for subjects}

\item{id}{The id of subjects}

\item{dist}{The distribution of the secondary outcomes (so far only allow "gaussian")}
}
\value{
A vector of information borrow scores
}
\description{
This function can be used to obtain the information borrow scores for the eMultiR estimate
}
