% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multir.R
\name{multir.est}
\alias{multir.est}
\title{The MultiR estimate}
\usage{
multir.est(Prop_multir, trt_ind, main_outcome)
}
\arguments{
\item{Prop_multir}{The calibration score}

\item{trt_ind}{Treatment assignment (eg., 0,1)}

\item{main_outcome}{The main outcome in binary}
}
\value{
A vector of estimates
}
\description{
This function can be used to obtain the MultiR estimate.
}
