% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multir.R
\name{multir.propensity.or}
\alias{multir.propensity.or}
\title{Calibration scores}
\usage{
multir.propensity.or(prob_fit1, prob_fit2,or_fit11, or_fit12,
or_fit01,or_fit02, trt_ind)
}
\arguments{
\item{prob_fit1}{The fitted probability in the treatment assignment model1}

\item{prob_fit2}{The fitted probability in the treatment assignment model2}

\item{or_fit11}{The estimating function for the CM model 1 for the treatment group}

\item{or_fit12}{The estimating function for the CM model 2 for the treatment group}

\item{or_fit01}{The estimating function for the CM model 1 for the control group}

\item{or_fit02}{The estimating function for the CM model 2 for the control group}

\item{trt_ind}{Treatment assignment (eg., 0,1)}
}
\value{
A vector of calibration scores
}
\description{
This function can be used to obtain the calibration scores for the MultiR estimate
}
